requirements.txt
cov-core==1.15.0
coverage==4.1
nose==1.3.7
nose-cov==1.6
pyserial==3.0.1
